requirements.txt
# Основные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9  # Драйвер PostgreSQL
python-jose[cryptography]==3.3.0  # JWT токены
passlib[bcrypt]==1.7.4  # Хеширование паролей
httpx==0.25.1  # HTTP клиент для API Gateway

# Дополнительные утилиты
alembic==1.12.1  # Миграции базы данных (опционально)
pydantic==2.5.0
python-dateutil==2.8.2  # Работа с датами
pytz==2023.3  # Часовые пояса

# RabbitMQ (опционально)
# pika==1.3.2  # Раскомментировать если нужен RabbitMQ

# Для логирования
structlog==23.1.0  # Структурированное логирование

# Для документации (опционально)
pygments==2.16.1  # Подсветка синтаксиса в документации
